Add function debounce
