+++ js/filter.js
@@ -14,13 +14,10 @@ const compareCommentsCount = (pictureA, pictureB) => pictureB.comments.length - 
 
 const getFilteredPictures = () => {
   switch (currentFilter) {
-  // case Filter.DEFAULT:
-    // return pictures;
     case Filter.RANDOM:
       return pictures.sort(sortRandomly).slice(0, PICTURES__COUNT);
     case Filter.DISCUSSED:
       return [...pictures].sort(compareCommentsCount);
-    //case Filter.DEFAULT:
     default:
       return [...pictures];
   }
@@ -31,8 +28,6 @@ const setOnFilterClick = (cb) => {
     if (!evt.target.classList.contains('img-filters__button')) {
       return;
     }
-
-    //filterElement.querySelector('img-filters__button--active').classList.remove('img-filters__button--active');
 
     filterElement.querySelectorAll('.img-filters__button').forEach((button) => {
       button.classList.remove('img-filters__button--active');
@@ -48,52 +43,6 @@ const init = (loaderPictures, cb) => {
   filterElement.classList.remove('img-filters--inactive');
   pictures = [...loaderPictures];
   setOnFilterClick(cb);
-  //cb(getFilteredPictures());
 };
 
-/* Функция взята из интернета и доработана
-// Источник - https://www.freecodecamp.org/news/javascript-debounce-example
-
-function debounce (callback, timeoutDelay = 500) {
-  // Используем замыкания, чтобы id таймаута у нас навсегда приклеился
-  // к возвращаемой функции с setTimeout, тогда мы его сможем перезаписывать
-  let timeoutId;
-
-  return (...rest) => {
-    // Перед каждым новым вызовом удаляем предыдущий таймаут,
-    // чтобы они не накапливались
-    clearTimeout(timeoutId);
-
-    // Затем устанавливаем новый таймаут с вызовом колбэка на ту же задержку
-    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-
-    // Таким образом цикл «поставить таймаут - удалить таймаут» будет выполняться,
-    // пока действие совершается чаще, чем переданная задержка timeoutDelay
-  };
-}
-
-// Функция взята из интернета и доработана
-// Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_throttle
-
-function throttle (callback, delayBetweenFrames) {
-  // Используем замыкания, чтобы время "последнего кадра" навсегда приклеилось
-  // к возвращаемой функции с условием, тогда мы его сможем перезаписывать
-  let lastTime = 0;
-
-  return (...rest) => {
-    // Получаем текущую дату в миллисекундах,
-    // чтобы можно было в дальнейшем
-    // вычислять разницу между кадрами
-    const now = new Date();
-
-    // Если время между кадрами больше задержки,
-    // вызываем наш колбэк и перезаписываем lastTime
-    // временем "последнего кадра"
-    if (now - lastTime >= delayBetweenFrames) {
-      callback.apply(this, rest);
-      lastTime = now;
-    }
-  };
-}
-*/
 export {init, getFilteredPictures};
