+++ js/main.js
@@ -1,15 +1,19 @@
 import { renderPictures } from './picture-modal.js';
 import { setUserFormSubmit} from './form.js';
 import { getData} from './api.js';
-import { showAlert} from './util.js';
+import { showAlert, debounce} from './util.js';
 import { showSuccessMessage } from './success-message.js';
 import { showErrorMessage } from './error-message.js';
 import { getFilteredPictures, init } from './filter.js';
 
 getData()
   .then((loaderPictures) => {
-    init(loaderPictures, (filteredPictures) => {
-      renderPictures(filteredPictures);
+    const debounceRenderPictures = debounce((pictures) => {
+      renderPictures(pictures);
+    });
+    init(loaderPictures, () => {
+      const filteredPictures = getFilteredPictures();
+      debounceRenderPictures(filteredPictures);
     });
     renderPictures(getFilteredPictures());
   })
