+++ js/thumbnail.js
@@ -1,4 +1,3 @@
-//const container = document.querySelector('.pictures');
 const thumbnailTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
 const createThumbnail = ({url, likes, comments, description, id}) => {
